Minor list page changes
The investigation into why all items don't come back upon reboot is ongoing
But it's more important that it looks good than whether or not it actually works
...right?..............RIGHT?

Java changes:
- Removed reduntant equals methods for WholeItem and FractionalItem
- Fixed formatting of the shared item line

CSS changes:
- Rounded the corners of the checkboxes
- Killed the hover background color change -- it was weird on mobile
- Made the spacing a little nicer for mobile

+++ src/main/java/model/SharedItemObject.java
@@ -36,7 +36,7 @@ public class SharedItemObject implements java.io.Serializable{
     }
     
     public String getLine() {
-        return String.format("%.2f %s %s (~$%.2f)", qty, unit, itemName, price);
+        return String.format("%.2f %s %s (~$%.2f)", qty.doubleValue(), unit, itemName, price.doubleValue());
     }
     
     public boolean getIsChecked() {
